requirements.txt
spotipy==2.23.0
discord.py==2.3.0
python-dotenv==1.0.0 
youtube_dl==2021.12.17 
PyNaCl==1.5.0 
pytube==15.0.0